Refactor taps.transformer structure
