+++ taps/transformer/__init__.py
@@ -1,10 +1,10 @@
 from __future__ import annotations
 
-from taps.transformer.config import TransformerConfig
 from taps.transformer.file import PickleFileTransformer
 from taps.transformer.file import PickleFileTransformerConfig
 from taps.transformer.null import NullTransformer
 from taps.transformer.null import NullTransformerConfig
 from taps.transformer.protocol import Transformer
+from taps.transformer.protocol import TransformerConfig
 from taps.transformer.proxy import ProxyTransformer
 from taps.transformer.proxy import ProxyTransformerConfig
