+++ taps/transformer/protocol.py
@@ -1,9 +1,14 @@
 from __future__ import annotations
 
+import abc
 from typing import Any
 from typing import Protocol
 from typing import runtime_checkable
 from typing import TypeVar
+
+from pydantic import BaseModel
+from pydantic import ConfigDict
+from pydantic import Field
 
 K = TypeVar('K')
 T = TypeVar('T')
@@ -46,4 +51,21 @@ class Transformer(Protocol[IdentifierT]):
         Returns:
             The resolved object.
         """
+        ...
+
+
+class TransformerConfig(BaseModel, abc.ABC):
+    """Abstract transformer configuration."""
+
+    name: str = Field(description='name of transformer type')
+
+    model_config: ConfigDict = ConfigDict(  # type: ignore[misc]
+        extra='ignore',
+        validate_default=True,
+        validate_return=True,
+    )
+
+    @abc.abstractmethod
+    def get_transformer(self) -> Transformer[Any]:
+        """Create a transformer from the configuration."""
         ...
