+++ taps/transformer/proxy.py
@@ -17,7 +17,7 @@ from pydantic import Field
 from pydantic import field_validator
 
 from taps.plugins import register
-from taps.transformer.config import TransformerConfig
+from taps.transformer.protocol import TransformerConfig
 
 T = TypeVar('T')
